fixed table refactor and did the main plot commands
