+++ DearPyGui/src/core/AppItems/plots/mvPlot.cpp
@@ -7,8 +7,57 @@
 #include "mvInput.h"
 #include "mvValueStorage.h"
 #include "mvGlobalIntepreterLock.h"
+#include "mvPythonExceptions.h"
 
 namespace Marvel {
+
+	// helpers
+	static bool CheckList(const char* plot, PyObject* list)
+	{
+		if (!PyList_Check(list))
+		{
+			ThrowPythonException(std::string(plot) + " add area series requires a list of floats.");
+			return false;
+		}
+		return true;
+	}
+
+	static bool CheckIfPlotOk(const char* name, mvAppItem* plot)
+	{
+		if (plot == nullptr)
+		{
+			ThrowPythonException(std::string(name) + " plot does not exist.");
+			return false;
+		}
+
+		if (plot->getType() != mvAppItemType::Plot)
+		{
+			ThrowPythonException(std::string(name) + " is not a plot.");
+			return false;
+		}
+		return true;
+	}
+
+	static bool Check2ArraySizes(const char* name, const std::vector<float>* first, const std::vector<float>* second)
+	{
+		if (second == nullptr)
+			return true;
+
+		return first->size() == second->size();
+	}
+
+	static bool CheckArraySizes(const char* name, std::vector<const std::vector<float>*> arrays)
+	{
+		for (size_t i = 0; i < arrays.size() - 1; i++)
+		{
+			if (!Check2ArraySizes(name, arrays[i], arrays[i + 1]))
+			{
+				ThrowPythonException(std::string(name) + " data list must be the same size.");
+				return false;
+			}
+		}
+		return true;
+	}
 
 	mvSeries::mvSeries(std::string name, const std::vector<const std::vector<float>*> data, ImPlotYAxis_ axis)
 		:
@@ -38,7 +87,7 @@ namespace Marvel {
 		{
 			if (x > m_maxX) m_maxX = x;
 			if (x < m_minX) m_minX = x;
-			
+
 		}
 
 		if (m_data.size() > 1)
@@ -53,7 +102,7 @@ namespace Marvel {
 	}
 
 	mvSeries::mvSeries(std::string name, const ImPlotPoint& boundsMin, const ImPlotPoint& boundsMax, ImPlotYAxis_ axis)
-		: 
+		:
 		m_name(std::move(name)),
 		m_axis(axis)
 	{
@@ -61,6 +110,137 @@ namespace Marvel {
 		m_maxY = (float)boundsMax.y;
 		m_minX = (float)boundsMin.x;
 		m_minY = (float)boundsMin.y;
+	}
+
+	void mvPlot::InsertParser(std::map<std::string, mvPythonParser>* parsers)
+	{
+		parsers->insert({ "add_plot", mvPythonParser({
+			{mvPythonDataType::String, "name"},
+			{mvPythonDataType::KeywordOnly},
+			{mvPythonDataType::String, "x_axis_name", "", "''"},
+			{mvPythonDataType::String, "y_axis_name", "", "''"},
+
+			// plot flags
+			{mvPythonDataType::Bool, "no_legend", "", "False"},
+			{mvPythonDataType::Bool, "no_menus", "", "False"},
+			{mvPythonDataType::Bool, "no_box_select", "", "False"},
+			{mvPythonDataType::Bool, "no_mouse_pos", "", "False"},
+			{mvPythonDataType::Bool, "no_highlight", "", "False"},
+			{mvPythonDataType::Bool, "no_child", "", "False"},
+			{mvPythonDataType::Bool, "query", "", "False"},
+			{mvPythonDataType::Bool, "crosshairs", "", "False"},
+			{mvPythonDataType::Bool, "anti_aliased", "", "False"},
+			{mvPythonDataType::Bool, "yaxis2", "", "False"},
+			{mvPythonDataType::Bool, "yaxis3", "", "False"},
+
+			// x axis flags
+			{mvPythonDataType::Bool, "xaxis_no_gridlines", "", "False"},
+			{mvPythonDataType::Bool, "xaxis_no_tick_marks", "", "False"},
+			{mvPythonDataType::Bool, "xaxis_no_tick_labels", "", "False"},
+			{mvPythonDataType::Bool, "xaxis_log_scale", "", "False"},
+			{mvPythonDataType::Bool, "xaxis_time", "", "False"},
+			{mvPythonDataType::Bool, "xaxis_invert", "", "False"},
+			{mvPythonDataType::Bool, "xaxis_lock_min", "", "False"},
+			{mvPythonDataType::Bool, "xaxis_lock_max", "", "False"},
+
+			// y axis flags
+			{mvPythonDataType::Bool, "yaxis_no_gridlines", "", "False"},
+			{mvPythonDataType::Bool, "yaxis_no_tick_marks", "", "False"},
+			{mvPythonDataType::Bool, "yaxis_no_tick_labels", "", "False"},
+			{mvPythonDataType::Bool, "yaxis_log_scale", "", "False"},
+			{mvPythonDataType::Bool, "yaxis_invert", "", "False"},
+			{mvPythonDataType::Bool, "yaxis_lock_min", "", "False"},
+			{mvPythonDataType::Bool, "yaxis_lock_max", "", "False"},
+
+			// y2 axis flags
+			{mvPythonDataType::Bool, "y2axis_no_gridlines", "", "False"},
+			{mvPythonDataType::Bool, "y2axis_no_tick_marks", "", "False"},
+			{mvPythonDataType::Bool, "y2axis_no_tick_labels", "", "False"},
+			{mvPythonDataType::Bool, "y2axis_log_scale", "", "False"},
+			{mvPythonDataType::Bool, "y2axis_invert", "", "False"},
+			{mvPythonDataType::Bool, "y2axis_lock_min", "", "False"},
+			{mvPythonDataType::Bool, "y2axis_lock_max", "", "False"},
+
+			// y3 axis flags
+			{mvPythonDataType::Bool, "y3axis_no_gridlines", "", "False"},
+			{mvPythonDataType::Bool, "y3axis_no_tick_marks", "", "False"},
+			{mvPythonDataType::Bool, "y3axis_no_tick_labels", "", "False"},
+			{mvPythonDataType::Bool, "y3axis_log_scale", "", "False"},
+			{mvPythonDataType::Bool, "y3axis_invert", "", "False"},
+			{mvPythonDataType::Bool, "y3axis_lock_min", "", "False"},
+			{mvPythonDataType::Bool, "y3axis_lock_max", "", "False"},
+
+			{mvPythonDataType::String, "parent", "Parent to add this item to. (runtime adding)", "''"},
+			{mvPythonDataType::String, "before", "This item will be displayed before the specified item in the parent. (runtime adding)", "''"},
+			{mvPythonDataType::Integer, "width", "", "-1"},
+			{mvPythonDataType::Integer, "height", "", "-1"},
+			{mvPythonDataType::Callable , "query_callback", "Callback ran when plot is queried. Should be of the form 'def Callback(sender, data)'\n Data is (x_min, x_max, y_min, y_max).", "None"},
+
+			{mvPythonDataType::Bool, "show_color_scale", "", "False"},
+			{mvPythonDataType::Float, "scale_min", "", "0.0"},
+			{mvPythonDataType::Float, "scale_max", "", "1.0"},
+			{mvPythonDataType::Integer, "scale_height", "", "100"},
+			{mvPythonDataType::String, "label", "", "''"},
+			{mvPythonDataType::Bool, "show", "Attempt to render", "True"},
+			{mvPythonDataType::Bool, "show_annotations", "", "True"},
+			{mvPythonDataType::Bool, "show_drag_lines", "", "True"},
+			{mvPythonDataType::Bool, "show_drag_points", "", "True"}
+
+		}, "Adds a plot widget.", "None", "Plotting") });
+
+		parsers->insert({ "add_drag_line", mvPythonParser({
+			{mvPythonDataType::String, "plot"},
+			{mvPythonDataType::String, "name"},
+			{mvPythonDataType::KeywordOnly},
+			{mvPythonDataType::String, "source", "", "''"},
+			{mvPythonDataType::FloatList, "color", "", "(0, 0, 0, -1)"},
+			{mvPythonDataType::Float, "thickness", "", "-1"},
+			{mvPythonDataType::Bool, "y_line", "", "False"},
+			{mvPythonDataType::Bool, "show_label", "", "True"},
+			{mvPythonDataType::Callable, "callback", "function to run when line is dragged", "None"},
+			{mvPythonDataType::Float, "default_value", "", "0.0"},
+		}, "Adds a drag line to a plot.", "None", "Plotting") });
+
+		parsers->insert({ "delete_drag_line", mvPythonParser({
+			{mvPythonDataType::String, "plot"},
+			{mvPythonDataType::String, "name"}
+		}, "Deletes a drag line if it exists.", "None", "Plotting") });
+
+		parsers->insert({ "add_drag_point", mvPythonParser({
+			{mvPythonDataType::String, "plot"},
+			{mvPythonDataType::String, "name"},
+			{mvPythonDataType::KeywordOnly},
+			{mvPythonDataType::String, "source", "", "''"},
+			{mvPythonDataType::FloatList, "color", "", "(0, 0, 0, -1)"},
+			{mvPythonDataType::Float, "radius", "", "4.0"},
+			{mvPythonDataType::Bool, "show_label", "", "True"},
+			{mvPythonDataType::Callable, "callback", "function to run when point is moved", "None"},
+			{mvPythonDataType::Float, "default_x", "", "0.0"},
+			{mvPythonDataType::Float, "default_y", "", "0.0"},
+		}, "Adds a drag point to a plot.", "None", "Plotting") });
+
+		parsers->insert({ "delete_drag_point", mvPythonParser({
+			{mvPythonDataType::String, "plot"},
+			{mvPythonDataType::String, "name"}
+		}, "Deletes a drag point if it exists.", "None", "Plotting") });
+
+		parsers->insert({ "add_annotation", mvPythonParser({
+			{mvPythonDataType::String, "plot"},
+			{mvPythonDataType::String, "text"},
+			{mvPythonDataType::Double, "x"},
+			{mvPythonDataType::Double, "y"},
+			{mvPythonDataType::Float, "xoffset"},
+			{mvPythonDataType::Float, "yoffset"},
+			{mvPythonDataType::KeywordOnly},
+			{mvPythonDataType::FloatList, "color", "", "(0, 0, 0, -1)"},
+			{mvPythonDataType::Bool, "clamped", "", "True"},
+			{mvPythonDataType::String, "tag", "", "''"},
+		}, "Adds an annotation to a plot.", "None", "Plotting") });
+
+		parsers->insert({ "delete_annotation", mvPythonParser({
+			{mvPythonDataType::String, "plot"},
+			{mvPythonDataType::String, "name"},
+		}, "Deletes an annotation", "None", "Plotting") });
 	}
 
 	mvPlot::mvPlot(const std::string& name, PyObject* queryCallback)
@@ -780,4 +960,346 @@ namespace Marvel {
 		checkbitset("y3axis_lock_min",       ImPlotAxisFlags_LockMin,      m_y3flags);
 		checkbitset("y3axis_lock_max",       ImPlotAxisFlags_LockMax,      m_y3flags);
 	}
+
+	PyObject* add_plot(PyObject* self, PyObject* args, PyObject* kwargs)
+	{
+		const char* name;
+		const char* xAxisName = "";
+		const char* yAxisName = "";
+
+		// plot flags
+		int no_legend = false;
+		int no_menus = false;
+		int no_box_select = false;
+		int no_mouse_pos = false;
+		int no_highlight = false;
+		int no_child = false;
+		int query = false;
+		int crosshairs = false;
+		int antialiased = false;
+		int yaxis2 = false;
+		int yaxis3 = false;
+
+		// x axis flags
+		int xaxis_no_gridlines = false;
+		int xaxis_no_tick_marks = false;
+		int xaxis_no_tick_labels = false;
+		int xaxis_log_scale = false;
+		int xaxis_time = false;
+		int xaxis_invert = false;
+		int xaxis_lock_min = false;
+		int xaxis_lock_max = false;
+
+		// y axis flags
+		int yaxis_no_gridlines = false;
+		int yaxis_no_tick_marks = false;
+		int yaxis_no_tick_labels = false;
+		int yaxis_log_scale = false;
+		int yaxis_invert = false;
+		int yaxis_lock_min = false;
+		int yaxis_lock_max = false;
+
+		// y2 axis flags
+		int y2axis_no_gridlines = false;
+		int y2axis_no_tick_marks = false;
+		int y2axis_no_tick_labels = false;
+		int y2axis_log_scale = false;
+		int y2axis_invert = false;
+		int y2axis_lock_min = false;
+		int y2axis_lock_max = false;
+
+		// y3 axis flags
+		int y3axis_no_gridlines = false;
+		int y3axis_no_tick_marks = false;
+		int y3axis_no_tick_labels = false;
+		int y3axis_log_scale = false;
+		int y3axis_invert = false;
+		int y3axis_lock_min = false;
+		int y3axis_lock_max = false;
+
+		const char* parent = "";
+		const char* before = "";
+		int width = -1;
+		int height = -1;
+		PyObject* query_callback = nullptr;
+
+		int show_color_scale = false;
+		float scale_min = 0.0f;
+		float scale_max = 1.0f;
+		int scale_height = 100;
+
+		const char* label = "";
+		int show = true;
+		int show_annotations = true;
+		int show_drag_lines = true;
+		int show_drag_points = true;
+
+		if (!(*mvApp::GetApp()->getParsers())["add_plot"].parse(args, kwargs, __FUNCTION__, &name, &xAxisName, &yAxisName,
+			&no_legend, &no_menus, &no_box_select, &no_mouse_pos, &no_highlight, &no_child, &query, &crosshairs, &antialiased,
+			&yaxis2, &yaxis3,
+			&xaxis_no_gridlines,
+			&xaxis_no_tick_marks,
+			&xaxis_no_tick_labels,
+			&xaxis_log_scale,
+			&xaxis_time,
+			&xaxis_invert,
+			&xaxis_lock_min,
+			&xaxis_lock_max,
+			&yaxis_no_gridlines,
+			&yaxis_no_tick_marks,
+			&yaxis_no_tick_labels,
+			&yaxis_log_scale,
+			&yaxis_invert,
+			&yaxis_lock_min,
+			&yaxis_lock_max,
+			&y2axis_no_gridlines,
+			&y2axis_no_tick_marks,
+			&y2axis_no_tick_labels,
+			&y2axis_log_scale,
+			&y2axis_invert,
+			&y2axis_lock_min,
+			&y2axis_lock_max,
+			&yaxis_no_gridlines,
+			&y3axis_no_tick_marks,
+			&y3axis_no_tick_labels,
+			&y3axis_log_scale,
+			&y3axis_invert,
+			&y3axis_lock_min,
+			&y3axis_lock_max,
+			&parent, &before, &width, &height, &query_callback, &show_color_scale, &scale_min, &scale_max,
+			&scale_height, &label, &show, &show_annotations, &show_drag_lines, &show_drag_points))
+			return ToPyBool(false);
+
+		if (query_callback)
+			Py_XINCREF(query_callback);
+
+		auto item = CreateRef<mvPlot>(name, query_callback);
+
+		item->checkConfigDict(kwargs);
+		item->setConfigDict(kwargs);
+		item->setExtraConfigDict(kwargs);
+
+		return ToPyBool(mvApp::GetApp()->getItemRegistry().addItemWithRuntimeChecks(item, parent, before));
+	}
+
+	PyObject* add_drag_line(PyObject* self, PyObject* args, PyObject* kwargs)
+	{
+		const char* plot;
+		const char* name;
+		const char* source = "";
+		PyObject* color = PyTuple_New(4);
+		PyTuple_SetItem(color, 0, PyLong_FromLong(0));
+		PyTuple_SetItem(color, 1, PyLong_FromLong(0));
+		PyTuple_SetItem(color, 2, PyLong_FromLong(0));
+		PyTuple_SetItem(color, 3, PyLong_FromLong(-1));
+		float thickness = 1.0f;
+		int y_line = false;
+		int show_label = true;
+		PyObject* callback = nullptr;
+		float default_value = 0.0f;
+
+		if (!(*mvApp::GetApp()->getParsers())["add_drag_line"].parse(args, kwargs, __FUNCTION__,
+			&plot, &name, &source, &color, &thickness, &y_line, &show_label, &callback, &default_value))
+			return GetPyNone();
+
+		auto aplot = mvApp::GetApp()->getItemRegistry().getItem(plot);
+
+		if (aplot == nullptr)
+		{
+			std::string message = plot;
+			ThrowPythonException(message + " plot does not exist.");
+			return GetPyNone();
+		}
+
+		if (aplot->getType() != mvAppItemType::Plot)
+		{
+			std::string message = plot;
+			ThrowPythonException(message + " is not a plot.");
+			return GetPyNone();
+		}
+
+		if (callback)
+			Py_XINCREF(callback);
+
+		mvPlot* graph = static_cast<mvPlot*>(aplot.get());
+		graph->updateDragLine(name, show_label, ToColor(color), thickness, y_line, callback, default_value, source);
+		return GetPyNone();
+	}
+
+	PyObject* delete_drag_line(PyObject* self, PyObject* args, PyObject* kwargs)
+	{
+		const char* plot;
+		const char* name;
+
+		if (!(*mvApp::GetApp()->getParsers())["delete_drag_line"].parse(args, kwargs, __FUNCTION__,
+			&plot, &name))
+			return GetPyNone();
+
+		auto aplot = mvApp::GetApp()->getItemRegistry().getItem(plot);
+
+		if (aplot == nullptr)
+		{
+			std::string message = plot;
+			ThrowPythonException(message + " plot does not exist.");
+			return GetPyNone();
+		}
+
+		if (aplot->getType() != mvAppItemType::Plot)
+		{
+			std::string message = plot;
+			ThrowPythonException(message + " is not a plot.");
+			return GetPyNone();
+		}
+
+		mvPlot* graph = static_cast<mvPlot*>(aplot.get());
+		graph->deleteDragLine(name);
+		return GetPyNone();
+	}
+
+	PyObject* add_drag_point(PyObject* self, PyObject* args, PyObject* kwargs)
+	{
+		const char* plot;
+		const char* name;
+		const char* source = "";
+		PyObject* color = PyTuple_New(4);
+		PyTuple_SetItem(color, 0, PyLong_FromLong(0));
+		PyTuple_SetItem(color, 1, PyLong_FromLong(0));
+		PyTuple_SetItem(color, 2, PyLong_FromLong(0));
+		PyTuple_SetItem(color, 3, PyLong_FromLong(-1));
+		float radius = 4.0f;
+		int show_label = true;
+		PyObject* callback = nullptr;
+		float default_x = 0.0f;
+		float default_y = 0.0f;
+
+		if (!(*mvApp::GetApp()->getParsers())["add_drag_point"].parse(args, kwargs, __FUNCTION__,
+			&plot, &name, &source, &color, &radius, &show_label, &callback, &default_x, &default_y))
+			return GetPyNone();
+
+		auto aplot = mvApp::GetApp()->getItemRegistry().getItem(plot);
+
+		if (aplot == nullptr)
+		{
+			std::string message = plot;
+			ThrowPythonException(message + " plot does not exist.");
+			return GetPyNone();
+		}
+
+		if (aplot->getType() != mvAppItemType::Plot)
+		{
+			std::string message = plot;
+			ThrowPythonException(message + " is not a plot.");
+			return GetPyNone();
+		}
+
+		if (callback)
+			Py_XINCREF(callback);
+
+		mvPlot* graph = static_cast<mvPlot*>(aplot.get());
+		double defaults[2] = { (double)default_x, (double)default_y };
+		graph->updateDragPoint(name, show_label, ToColor(color), radius, callback, defaults, source);
+		return GetPyNone();
+	}
+
+	PyObject* delete_drag_point(PyObject* self, PyObject* args, PyObject* kwargs)
+	{
+		const char* plot;
+		const char* name;
+
+		if (!(*mvApp::GetApp()->getParsers())["delete_drag_point"].parse(args, kwargs, __FUNCTION__,
+			&plot, &name))
+			return GetPyNone();
+
+		auto aplot = mvApp::GetApp()->getItemRegistry().getItem(plot);
+
+		if (aplot == nullptr)
+		{
+			std::string message = plot;
+			ThrowPythonException(message + " plot does not exist.");
+			return GetPyNone();
+		}
+
+		if (aplot->getType() != mvAppItemType::Plot)
+		{
+			std::string message = plot;
+			ThrowPythonException(message + " is not a plot.");
+			return GetPyNone();
+		}
+
+		mvPlot* graph = static_cast<mvPlot*>(aplot.get());
+		graph->deleteDragPoint(name);
+		return GetPyNone();
+	}
+
+	PyObject* add_annotation(PyObject* self, PyObject* args, PyObject* kwargs)
+	{
+		const char* plot;
+		const char* text;
+		double x;
+		double y;
+		float xoffset;
+		float yoffset;
+		PyObject* color = PyTuple_New(4);
+		PyTuple_SetItem(color, 0, PyLong_FromLong(0));
+		PyTuple_SetItem(color, 1, PyLong_FromLong(0));
+		PyTuple_SetItem(color, 2, PyLong_FromLong(0));
+		PyTuple_SetItem(color, 3, PyLong_FromLong(0));
+		int clamped = true;
+		const char* tag = "";
+
+		if (!(*mvApp::GetApp()->getParsers())["add_annotation"].parse(args, kwargs, __FUNCTION__,
+			&plot, &text, &x, &y, &xoffset, &yoffset, &color, &clamped, &tag))
+			return GetPyNone();
+
+		auto aplot = mvApp::GetApp()->getItemRegistry().getItem(plot);
+
+		if (aplot == nullptr)
+		{
+			std::string message = plot;
+			ThrowPythonException(message + " plot does not exist.");
+			return GetPyNone();
+		}
+
+		if (aplot->getType() != mvAppItemType::Plot)
+		{
+			std::string message = plot;
+			ThrowPythonException(message + " is not a plot.");
+			return GetPyNone();
+		}
+
+		mvPlot* graph = static_cast<mvPlot*>(aplot.get());
+		graph->updateAnnotation(tag, x, y, xoffset, yoffset, ToColor(color), text, clamped);
+		return GetPyNone();
+	}
+
+	PyObject* delete_annotation(PyObject* self, PyObject* args, PyObject* kwargs)
+	{
+		const char* plot;
+		const char* name;
+
+		if (!(*mvApp::GetApp()->getParsers())["delete_annotation"].parse(args, kwargs, __FUNCTION__,
+			&plot, &name))
+			return GetPyNone();
+
+		auto aplot = mvApp::GetApp()->getItemRegistry().getItem(plot);
+
+		if (aplot == nullptr)
+		{
+			std::string message = plot;
+			ThrowPythonException(message + " plot does not exist.");
+			return GetPyNone();
+		}
+
+		if (aplot->getType() != mvAppItemType::Plot)
+		{
+			std::string message = plot;
+			ThrowPythonException(message + " is not a plot.");
+			return GetPyNone();
+		}
+
+		mvPlot* graph = static_cast<mvPlot*>(aplot.get());
+		graph->deleteAnnotation(name);
+		return GetPyNone();
+	}
+
 }
